Converted all existing diagrams to use the new framework. (#24)
This converts the existing diagram code over to using the new framework for capturing diagrams.

Some of the output varies a little bit, because of differences in the pixel ratio and timing. The variations are minor, and most of the text and drawing looks a lot sharper than the old images.

Not to mention that what once took about 30 minutes and two passes, now takes 7 seconds.