Allows generation of just a subset of the diagrams by passing command line options. (#32)
Passing either --category or --name to generate.dart will restrict the generated diagrams to just those diagrams that match. The flags can be specified multiple times to specify more than one. The category is the directory name that the assets end up in, and the name can be the base name of the diagram, with or without the suffix.