Add owl.png for use with Image Widget samples
Originally, I generated a diagram using the 
Image Widget + owl.png, but this could only result
in an mp4 movie due to the asynchronous nature of
Image loading.

I didn’t think that made much sense to show a 
video in the API docs for an image Widget, so I’ve
simply included owl.png and will link to that 
From the docs.