Modernize Dart, update analysis options, add --device option (#116)
This PR mainly updates all of the assets-for-api-docs code to be "modern" Dart (no new operators, correct casting, etc.). I also updated the project files to be modern versions by running "flutter create" again to recreate the projects.

I updated the analysis_options.yaml file to match what was in the Flutter repository, and fixed all the issues. I didn't do NNBD changes yet, but that would be desirable.

I also added support for a "--device" argument, since now that we have desktop, it wasn't possible to run the generator, since you couldn't tell flutter which device to run on.

It still supports running on Android, and I (sort of) got it running on Linux if you pass --device=linux, but I think that needs some work still, since you can't pass command line arguments to a desktop app yet (so it always builds ALL of the assets).

I also pinned the CI to using the dev branch of flutter.