Convert to null safety, update application shims (#127)
This converts the diagram generator to null safety, updates the application shims to the latest version, and updates the generated images to images generated with the new code.

Not all dependent packages are null safe yet (args, path_provider, etc), so I had to add --no-sound-null-safety in a few places.

I swapped out the version of FakeProcessManager for the one in the null safe version of process_runner and got rid of the dependency on Mockito.

Re-running the script to regenerate the images caused the floating_action_button_location images to change, but that is because a bug was fixed without re-running it. The new images look much better.