Add the snippets package to the repo. (#143)
This adds a new package to this repo that allows edit source code samples. It can also create new samples on an existing Dart element.

Eventually, I would like to migrate the snippets tool in the repo to either use the snippets package from here, or to just use that package directly via pub global activate the way that we do dartdoc.

I have not published the snippets package yet, but I will once this PR is committed.

Much of the code in the snippets package started out as a copy of the code in the flutter repo's snippets tool and analyze_sample_code.dart. This combines them into a single package adds methods for creating new samples and parsing the Dart code with the Dart analyzer's AST parser.

This forms the basis for a follow-on PR that adds a Flutter app that can extract and reinsert samples so that they can be edited, debugged, and run in an IDE.