Add the Sampler app (#144)
This adds a new Flutter application called sampler to the repo.

The app allows the Flutter developer to edit sample code in the doc comments of a Dart file. It doesn't allow editing in the app itself, but rather launches an external IDE or file browser to allow editing of the sample app in your favorite environment, and then places the edited code back into the documentation comments in the original source file.

To launch the app, just run the bin/sampler script. It will detect your platform and run the right version.

This app only works on Linux, macOS, and Windows, since mobile OSs don't have access to the Flutter repo.