Sort imports automatically when creating snippets (#145)
This change adds a utility for sorting imports according to the rules that the analysis server uses. It repurposes analysis server code so that it doesn't need the analysis_server package (which currently requires sdk < 2.0.0).

This way we won't need to suppress the directives_ordering lint when doing snippet analysis, and the snippets will have proper import ordering, even if the template and the sample source both have imports.

It pulls comments associated with the import along with them, to keep the comments in proper order as well.

Will help with flutter/flutter#81697 once snippets are generated using this package.