Make snippets conform to the old snippets tool by formatting output (#156)
This makes the snippets tool output conform more closely with the output from the original one in the Flutter repo. This PR changes the output so that it is formatted with the Dart formatter, and so that it only outputs samples if they are dartpad or sample samples, not when they are snippets.