Move sampler compilation into a temporary directory. (#164)
On my previous change (#160) I was only able to test it locally via "pub global activate --source path sampler", which is using the source in place rather than installing it in the pub cache. Turns out that writing the build the pub cache did some weird things to pub, and Flutter refuses to run pub get inside of one (sensibly).

So, instead of doing the build in place, I'm now copying the code to a temp dir and building it there. Yes, this means that you'll have to recompile if you reboot, but that's not a huge price to pay, and you'll have to recompile if you update anyhow.

I'm not super happy with the performance of the startup the first time (it takes about a minute), but subsequent times are reasonable.

I also eliminated the platform directories and generate them in the temp dir with the flutter used to do the building.