Add additional text diagrams demonstrating overflow (#228)
Related to this PR: https://github.com/flutter/flutter/pull/139363

Note that in addition to the new diagrams I also changed the "text_ellipsis" diagram. Most notably by adding a border which I think makes it more clear how the text is cut off.

I also deleted the "text" diagram since I found it a bit superfluous and since it wasn't referred to in the text using the diagram in the documentation.